refactor: move all component-pages in separate folder
