+++ src/components/item/item.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/constants";
-@import "../../../styles/mixins";
+@import "../../styles/constants";
+@import "../../styles/mixins";
 
 .item {
   width: calc(50% - 15px);
